Conflicts:
	boards/atsamr21-xpro/Makefile.include
	boards/atsamr21-xpro/board.c
	boards/atsamr21-xpro/dist/flash.sh
	boards/atsamr21-xpro/include/board.h
	boards/atsamr21-xpro/include/periph_conf.h
	boards/atsamr21-xpro/system_samr21xxxA.c

+++ cpu/lpc1768/atom.c
@@ -114,7 +114,7 @@ NORETURN void sched_task_return(void)
 char * thread_stack_init(void *(*task_func)(void *), void *arg, void *stack_start, int stack_size)
 {
     unsigned int * stk;
-    stk = (unsigned int *) (stack_start + stack_size);
+    stk = (unsigned int *) ((uintptr_t) stack_start + stack_size);
 
     /* marker */
     stk--;
